pyproject.toml
[tool.poetry]
name = "murasaki"
version = "0.1.0"
description = ""
authors = ["ginstrom <r-ginstrom@mercari.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2.8"
pytest = "^7.4.3"
pytest-coverage = "^0.0"
django-ckeditor = "^6.7.0"
django-parler = "^2.3"
django-rosetta = "^0.10.0"
psycopg = "^3.1.16"
pytest-django = "^4.7.0"
coverage = "^7.4.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"
pillow = "^10.1.0"
django-imagekit = "^5.0.0"
django-embed-video = "^1.4.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
